chart.js continues duplicate layers
