+++ src/stat-colors.js
@@ -1,6 +1,5 @@
 // stat-colors.js
 
-import {downloaded} from './mock-data.js';
 import Chart from 'chart.js';
 import ChartDataLabels from 'chartjs-plugin-datalabels';
 
@@ -28,67 +27,61 @@ const getColorsStat = (tasks) => {
   return {colors, quantites};
 };
 
-const colorWrap = document.querySelector(`.statistic__colors-wrap`);
-const colorsCtx = colorWrap.querySelector(`.statistic__colors`);
+const colorsWrap = document.querySelector(`.statistic__colors-wrap`);
+const colorsCtx = colorsWrap.querySelector(`.statistic__colors`);
 
-const activateColorsStat = () => {
-  colorWrap.classList.remove(`visually-hidden`);
-
-  const colorsChart = new Chart(colorsCtx, {
-    plugins: [ChartDataLabels],
-    type: `pie`,
-    data: {
-      labels: getColorsStat(downloaded.tasks).colors,
-      datasets: [{
-        data: getColorsStat(downloaded.tasks).quantites,
-        backgroundColor: [`#ff3cb9`, `#ffe125`, `#0c5cdd`, `#000000`, `#31b55c`]
-      }]
+const getColorsChart = () => new Chart(colorsCtx, {
+  plugins: [ChartDataLabels],
+  type: `pie`,
+  data: {
+    labels: [],
+    datasets: [{
+      data: [],
+      backgroundColor: []
+    }]
+  },
+  options: {
+    plugins: {
+      datalabels: {
+        display: false
+      }
     },
-    options: {
-      plugins: {
-        datalabels: {
-          display: false
+    tooltips: {
+      callbacks: {
+        label: (tooltipItem, data) => {
+          const allData = data.datasets[tooltipItem.datasetIndex].data;
+          const tooltipData = allData[tooltipItem.index];
+          const total = allData.reduce((acc, it) => acc + parseFloat(it));
+          const tooltipPercentage = Math.round((tooltipData / total) * 100);
+          return `${tooltipData} TASKS — ${tooltipPercentage}%`;
         }
       },
-      tooltips: {
-        callbacks: {
-          label: (tooltipItem, data) => {
-            const allData = data.datasets[tooltipItem.datasetIndex].data;
-            const tooltipData = allData[tooltipItem.index];
-            const total = allData.reduce((acc, it) => acc + parseFloat(it));
-            const tooltipPercentage = Math.round((tooltipData / total) * 100);
-            return `${tooltipData} TASKS — ${tooltipPercentage}%`;
-          }
-        },
-        displayColors: false,
-        backgroundColor: `#ffffff`,
-        bodyFontColor: `#000000`,
-        borderColor: `#000000`,
-        borderWidth: 1,
-        cornerRadius: 0,
-        xPadding: 15,
-        yPadding: 15
-      },
-      title: {
-        display: true,
-        text: `DONE BY: COLORS`,
-        fontSize: 16,
-        fontColor: `#000000`
-      },
-      legend: {
-        position: `left`,
-        labels: {
-          boxWidth: 15,
-          padding: 25,
-          fontStyle: 500,
-          fontColor: `#000000`,
-          fontSize: 13
-        }
+      displayColors: false,
+      backgroundColor: `#ffffff`,
+      bodyFontColor: `#000000`,
+      borderColor: `#000000`,
+      borderWidth: 1,
+      cornerRadius: 0,
+      xPadding: 15,
+      yPadding: 15
+    },
+    title: {
+      display: true,
+      text: `DONE BY: COLORS`,
+      fontSize: 16,
+      fontColor: `#000000`
+    },
+    legend: {
+      position: `left`,
+      labels: {
+        boxWidth: 15,
+        padding: 25,
+        fontStyle: 500,
+        fontColor: `#000000`,
+        fontSize: 13
       }
     }
-  });
+  }
+});
 
-  colorsChart.update();
-};
-
-export default activateColorsStat;
+export {getColorsStat, getColorsChart, colorsWrap};
