+++ src/stat-tags.js
@@ -1,6 +1,5 @@
 // stat-tags.js
 
-import {downloaded} from './mock-data.js';
 import Chart from 'chart.js';
 import ChartDataLabels from 'chartjs-plugin-datalabels';
 
@@ -33,66 +32,58 @@ const getTagsStat = (tasks) => {
 const tagsWrap = document.querySelector(`.statistic__tags-wrap`);
 const tagsCtx = tagsWrap.querySelector(`.statistic__tags`);
 
-const getNewChart = () => {
-  return new Chart(tagsCtx, {
-    plugins: [ChartDataLabels],
-    type: `pie`,
-    data: {
-      labels: getTagsStat(downloaded.tasks).names,
-      datasets: [{
-        data: getTagsStat(downloaded.tasks).quantites,
-        backgroundColor: [`#ff3cb9`, `#ffe125`, `#0c5cdd`, `#000000`, `#31b55c`]
-      }]
+const getTagsChart = () => new Chart(tagsCtx, {
+  plugins: [ChartDataLabels],
+  type: `pie`,
+  data: {
+    labels: [],
+    datasets: [{
+      data: [],
+      backgroundColor: [`#ff3cb9`, `#ffe125`, `#0c5cdd`, `#000000`, `#31b55c`]
+    }]
+  },
+  options: {
+    plugins: {
+      datalabels: {
+        display: false
+      }
     },
-    options: {
-      plugins: {
-        datalabels: {
-          display: false
+    tooltips: {
+      callbacks: {
+        label: (tooltipItem, data) => {
+          const allData = data.datasets[tooltipItem.datasetIndex].data;
+          const tooltipData = allData[tooltipItem.index];
+          const total = allData.reduce((acc, it) => acc + parseFloat(it));
+          const tooltipPercentage = Math.round((tooltipData / total) * 100);
+          return `${tooltipData} TASKS — ${tooltipPercentage}%`;
         }
       },
-      tooltips: {
-        callbacks: {
-          label: (tooltipItem, data) => {
-            const allData = data.datasets[tooltipItem.datasetIndex].data;
-            const tooltipData = allData[tooltipItem.index];
-            const total = allData.reduce((acc, it) => acc + parseFloat(it));
-            const tooltipPercentage = Math.round((tooltipData / total) * 100);
-            return `${tooltipData} TASKS — ${tooltipPercentage}%`;
-          }
-        },
-        displayColors: false,
-        backgroundColor: `#ffffff`,
-        bodyFontColor: `#000000`,
-        borderColor: `#000000`,
-        borderWidth: 1,
-        cornerRadius: 0,
-        xPadding: 15,
-        yPadding: 15
-      },
-      title: {
-        display: true,
-        text: `DONE BY: TAGS`,
-        fontSize: 16,
-        fontColor: `#000000`
-      },
-      legend: {
-        position: `left`,
-        labels: {
-          boxWidth: 15,
-          padding: 25,
-          fontStyle: 500,
-          fontColor: `#000000`,
-          fontSize: 13
-        }
+      displayColors: false,
+      backgroundColor: `#ffffff`,
+      bodyFontColor: `#000000`,
+      borderColor: `#000000`,
+      borderWidth: 1,
+      cornerRadius: 0,
+      xPadding: 15,
+      yPadding: 15
+    },
+    title: {
+      display: true,
+      text: `DONE BY: TAGS`,
+      fontSize: 16,
+      fontColor: `#000000`
+    },
+    legend: {
+      position: `left`,
+      labels: {
+        boxWidth: 15,
+        padding: 25,
+        fontStyle: 500,
+        fontColor: `#000000`,
+        fontSize: 13
       }
     }
-  });
-};
+  }
+});
 
-const activateTagsStat = () => {
-  tagsWrap.classList.remove(`visually-hidden`);
-  const tagsChart = getNewChart();
-  tagsChart.update();
-};
-
-export default activateTagsStat;
+export {getTagsStat, getTagsChart, tagsWrap};
